+++ main-server/src/config.js
@@ -276,9 +276,13 @@ export const prometheusHttpsCertPath =
     ? process.env.PROMETHEUS_HTTPS_CERT_PATH
     : path.join(__dirname, '..', 'dev_https_key', 'cert.pem');
 
-export const telemetryLoggingEnabled =
-  process.env.ENABLE_TELEMETRY_LOGGING === 'true';
-export const telemetryDbIp = process.env.TELEMETRY_DB_IP || dbIp;
+export const telemetryLoggingEnabled = process.env.ENABLE_TELEMETRY_LOGGING
+  ? process.env.ENABLE_TELEMETRY_LOGGING === 'true'
+  : true;
+export const telemetryDbHost =
+  (process.env.TELEMETRY_DB_HOST
+    ? process.env.TELEMETRY_DB_HOST
+    : process.env.TELEMETRY_DB_IP) || dbIp;
 export const telemetryDbPort = process.env.TELEMETRY_DB_PORT || dbPort;
 export const telemetryDbPassword =
   process.env.TELEMETRY_DB_PASSWORD || dbPassword;