+++ main-server/src/env_var_validate.js
@@ -167,11 +167,3 @@ if (process.env.CALLBACK_API_VERSION != null) {
     process.exit(1);
   }
 }
-
-if (process.env.ENABLE_TELEMETRY_LOGGING == null) {
-  console.error(
-    'ERROR:',
-    '"ENABLE_TELEMETRY_LOGGING" environment variable is not set. Process will now exit.'
-  );
-  process.exit(1);
-}